assignment/backend/requirements.txt
fastapi[standard]
python-keycloak
